+++ kaye/cli/cli_main.py
@@ -20,6 +20,3 @@ cli_parser.set_defaults(func=_cli_main)
 cli_subparser = cli_parser.add_subparsers(title="subcommands")
 
 register_cli_prompt_parser(cli_subparser)
-
-
-# todo write tests for cli